Better test script

+++ hack/test-cwnp.sh
@@ -1,26 +1,57 @@
 #!/bin/bash
+
+# Create ns where all CRD ClusterwideNetworkPolicy reside
 kubectl create ns firewall
+
+# Apply CRD ClusterwideNetworkPolicy to local service-cluster
 kubectl apply -f https://raw.githubusercontent.com/metal-stack/firewall-controller/master/config/crd/bases/metal-stack.io_clusterwidenetworkpolicies.yaml
+
+# Apply CRD Postgres to control-plane-cluster corresponding to kubeconfig
 kubectl --kubeconfig kubeconfig apply -f config/samples/_test_cwnp.yaml
-until kubectl get clusterwidenetworkpolicy -n firewall -o jsonpath='{.items[0].metadata.name}' > /dev/null 2>&1; do
-    echo 'waiting for cwnp'
+
+# Wait till CRD ClusterwideNetworkPolicy is available
+until kubectl get clusterwidenetworkpolicy -n firewall -o jsonpath='{.items[0]}' > /dev/null 2>&1; do
+    echo 'waiting for CRD ClusterwideNetworkPolicy'
     sleep 1s
 done
-echo 'cwnp is ready'
-PG_PROJECT_ID=$(kubectl --kubeconfig kubeconfig get postgres -A -o jsonpath='{.items[0].spec.projectID}')
-PG_UID=$(kubectl --kubeconfig kubeconfig get postgres -A -o jsonpath='{.items[0].metadata.uid}')
-POLICY_NAME=$(kubectl get clusterwidenetworkpolicy -n firewall -o jsonpath='{.items[0].metadata.name}')
-echo -e "===> creation test \c"
-if [ "$POLICY_NAME" == "$PG_PROJECT_ID--$PG_UID" ]; then
-    echo 'passed'
-else
-    echo 'failed'
+
+echo -e 'CRD ClusterwideNetworkPolicy creation test \c'
+
+# Check if only one CRD ClusterwideNetworkPolicy is creatd
+if kubectl get clusterwidenetworkpolicy -n firewall -o jsonpath='{.items[1]}' > /dev/null 2>&1 ; then
+    echo 'failed: more than one CRD ClusterwideNetworkPolicy'
     exit 1
 fi
+
+# Check if the name is correct
+POLICY_NAME=$(kubectl get clusterwidenetworkpolicy -n firewall -o jsonpath='{.items[0].metadata.name}')
+PG_PROJECT_ID=$(kubectl --kubeconfig kubeconfig get postgres -A -o jsonpath='{.items[0].spec.projectID}')
+PG_UID=$(kubectl --kubeconfig kubeconfig get postgres -A -o jsonpath='{.items[0].metadata.uid}')
+if [ "$POLICY_NAME" != "$PG_PROJECT_ID--$PG_UID" ]; then
+    echo 'failed: wrong name'
+    exit 1
+fi
+
+
+# Check the CIDR was correctly transferred
+POSTGRES_CIDR=$(kubectl --kubeconfig kubeconfig get postgres -A -o jsonpath='{.items[0].spec.accessList.sourceRanges[0]}')
+POLICY_CIDR=$(kubectl get clusterwidenetworkpolicy -A -o jsonpath='{.items[0].spec.ingress[0].from[0].cidr}')
+if [ "$POLICY_CIDR" != "$POSTGRES_CIDR" ]; then
+    echo 'failed: wrong CIDR'
+    exit 1
+fi
+
+echo 'passed'
+
+# Delete CRD ClusterwideNetworkPolicy
 kubectl --kubeconfig kubeconfig delete -f config/samples/_test_cwnp.yaml
+
+# Check if CRD ClusterwideNetworkPolicy is deleted
 while kubectl get clusterwidenetworkpolicy -n firewall "$POLICY_NAME" > /dev/null 2>&1; do
-    echo 'cwnp still there'
+    echo 'CRD ClusterwideNetworkPolicy still there'
     sleep 1s
 done
-echo "===> deletion test passed"
+echo 'CRD ClusterwideNetworkPolicy deletion test passed'
+
+# Clean up
 kubectl delete ns firewall
